requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord==1.0.1
discord.py==1.5.0
EasyProcess==0.3
entrypoint2==0.2.3
idna==2.10
MouseInfo==0.1.3
mss==6.0.0
multidict==4.7.6
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
pyscreenshot==2.2
PyScreeze==0.1.26
pytesseract==0.3.6
python-dotenv==0.14.0
PyTweening==1.0.3
yarl==1.5.1
